refactor: 픽셀값 수정
픽셀값 수정

(#303)
