+++ src/styles/NoticeListStyle.scss
@@ -6,7 +6,7 @@
 }
 .noticeListLayout {
   width: 100%;
-  height: 871px;
+  height: 98.53vh;
   background: $white100Color;
   display: flex;
   flex-direction: row;
@@ -17,7 +17,7 @@
     margin-right: 16.47vw;
 
     .noticeListHead {
-      margin-top: 17.2vh;
+      margin-top: 22.62vh;
       margin-bottom: 3.62vh;
       width: 34vw;
       height: 4.87vh;
@@ -67,8 +67,8 @@
   .noticeBoxContainer {
     width: 37.7vw;
     height: 64.48vh;
-    margin-top: 17.2vh;
-    margin-bottom: 10.18vh;
+    margin-top: 22.62vh;
+    margin-bottom: 22.62vh;
     display: grid;
     grid-template-columns: 17.2vw 17.2vw;
     grid-template-rows: 29.41vh 29.41vh;
@@ -135,5 +135,5 @@
 
 .fadeInAnimation {
   position: relative;
-  animation: fadeIn 2s ease-in-out;
+  animation: fadeIn 1.5s ease-in-out;
 }
