+++ src/styles/ThemeStyles.scss
@@ -16,6 +16,7 @@ $white100Color: rgba(255, 255, 255, 1);
 $errorColor: rgba(221, 82, 87, 1);
 $numberColor: rgba(53, 60, 73, 1);
 $hoverAndClickColor: rgba(77, 125, 220, 1);
+$footerColor: rgba(246, 249, 253, 1);
 
 @mixin displayFont($color) {
   color: $color;
